Add endpoint to get team count by date created

+++ api/src/main/java/com/ford/labs/retroquest/metrics/MetricsController.java
@@ -19,7 +19,7 @@ public class MetricsController {
     }
 
     @GetMapping("/team/count")
-    public int getTeamCount(
+    public long getTeamCount(
             @RequestParam(name = "start", required = false)
             @DateTimeFormat(iso = DateTimeFormat.ISO.DATE)
                     LocalDate startDate,
